Update tooling projects to more modern dependencies.
We really don't need to use netcoreapp1.0 anywhere in our tooling now.
We still have some desktop .NET references around. We may want to prune our tool set anyway at some point - several of these tools haven't been used in a long time.
