+++ tools/Google.Cloud.Tools.GenerateReleaseNotes/Google.Cloud.Tools.GenerateReleaseNotes.csproj
@@ -1,13 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <IsPackable>false</IsPackable>
     <Description>Console app to generate release notes from git logs</Description>
-    <TargetFramework>net451</TargetFramework>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
     <OutputType>Exe</OutputType>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="LibGit2Sharp" Version="0.24.0" />
+    <PackageReference Include="LibGit2Sharp" Version="0.25.0" />
   </ItemGroup>
 
   <ItemGroup>
