+++ tools/Google.Cloud.Tools.ShowReverseDependencies/Google.Cloud.Tools.ShowReverseDependencies.csproj
@@ -3,11 +3,8 @@
   <PropertyGroup>
     <IsPackable>false</IsPackable>
     <Description>Console app to show the reverse dependencies on NuGet packages</Description>
-    <TargetFramework>netcoreapp1.0</TargetFramework>
-    <AssemblyName>Google.Cloud.Tools.ShowReverseDependencies</AssemblyName>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
     <OutputType>Exe</OutputType>
-    <PackageId>Google.Cloud.Tools.ShowReverseDependencies</PackageId>
-    <RuntimeFrameworkVersion>1.0.4</RuntimeFrameworkVersion>
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
   </PropertyGroup>
