+++ tools/Google.Cloud.Tools.ValidateLinks/Google.Cloud.Tools.ValidateLinks.csproj
@@ -4,21 +4,13 @@
     <IsPackable>false</IsPackable>
     <Description>Console app to validate relative links in generated HTML</Description>
     <TargetFramework>net451</TargetFramework>
-    <AssemblyName>Google.Cloud.Tools.ValidateLinks</AssemblyName>
     <OutputType>Exe</OutputType>
-    <PackageId>Google.Cloud.Tools.ValidateLinks</PackageId>
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Google.Cloud.Tools.Common\Google.Cloud.Tools.Common.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="TidyHtml5Managed" Version="1.1.0" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net451' ">
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Xml" />
     <Reference Include="System" />
